Break Federator.Federate into a few modules
Also:
* Move Federator.PolysemyOrphans to Federator.Utils.PolysemyServerError

* Break Test.Federator.Federate into two modules
