+++ services/federator/test/unit/Main.hs
@@ -21,7 +21,8 @@ module Main
 where
 
 import Imports
-import qualified Test.Federator.Federate
+import qualified Test.Federator.ExternalServer
+import qualified Test.Federator.InternalServer
 import qualified Test.Federator.Options
 import Test.Tasty
 
@@ -31,5 +32,6 @@ main =
     testGroup
       "Tests"
       [ Test.Federator.Options.tests,
-        Test.Federator.Federate.tests
+        Test.Federator.InternalServer.tests,
+        Test.Federator.ExternalServer.tests
       ]
