+++ services/federator/test/unit/Test/Federator/InternalServer.hs
@@ -1,30 +1,12 @@
-{-# LANGUAGE PolyKinds #-}
-{-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE TypeApplications #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
--- This file is part of the Wire Server implementation.
---
--- Copyright (C) 2020 Wire Swiss GmbH <opensource@wire.com>
---
--- This program is free software: you can redistribute it and/or modify it under
--- the terms of the GNU Affero General Public License as published by the Free
--- Software Foundation, either version 3 of the License, or (at your option) any
--- later version.
---
--- This program is distributed in the hope that it will be useful, but WITHOUT
--- ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
--- FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
--- details.
---
--- You should have received a copy of the GNU Affero General Public License along
--- with this program. If not, see <https://www.gnu.org/licenses/>.
-
-module Test.Federator.Federate where
+module Test.Federator.InternalServer where
 
 import Data.Domain (Domain (Domain))
 import Data.Either.Validation
-import Federator.Federate
+import Federator.Discovery (LookupError (LookupErrorSrvNotAvailable))
+import Federator.InternalServer (callRemote)
+import Federator.Remote (Remote, RemoteError (RemoteErrorDiscoveryFailure))
 import Imports
 import Mu.GRpc.Client.Record
 import qualified Network.HTTP.Types as HTTP
@@ -37,7 +19,6 @@ import Test.Tasty.HUnit (assertBool, assertEqual, testCase)
 import Wire.API.Federation.GRPC.Types
 
 genMock ''Remote
-genMock ''Brig
 
 tests :: TestTree
 tests =
@@ -48,9 +29,6 @@ tests =
           remoteCallFailureErrCode,
           remoteCallFailureErrStr,
           remoteCallFailureErrConn
-        ],
-      testGroup "with local" $
-        [ localCallBrigSuccess
         ]
     ]
 
@@ -125,21 +103,6 @@ remoteCallFailureErrConn =
       let expectedCall = ValidatedRemoteCall (Domain validDomainText) validLocalPart
       embed $ assertEqual "one remote call should be made" [expectedCall] actualCalls
       embed $ assertBool "the response should have error" (isResponseError res)
-
--- TODO: Add more tests
-localCallBrigSuccess :: TestTree
-localCallBrigSuccess =
-  testCase "should sucessfully return on HTTP 200" $
-    runM . evalMock @Brig @IO $ do
-      mockBrigCallReturns @IO (\_ _ _ _ -> pure (HTTP.status200, Just "response body"))
-      let request = LocalCall Brig (HTTPMethod HTTP.GET) "/users" [QueryParam "handle" "foo"] mempty
-
-      res <- mock @Brig @IO $ callLocal request
-
-      actualCalls <- mockBrigCallCalls @IO
-      let expectedCall = (HTTP.GET, "/users", [QueryParam "handle" "foo"], mempty)
-      embed $ assertEqual "one call to brig should be made" [expectedCall] actualCalls
-      embed $ assertEqual "response should be success with correct body" (ResponseHTTPResponse (HTTPResponse 200 "response body")) res
 
 isResponseError :: Response -> Bool
 isResponseError (ResponseErr _) = True
